@@ -166,9 +166,11 @@ ensure_newline_before_comments = true
 skip_glob = ["*/grammars/*", "*/build/*"]
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 line-length = 88
 fix = true
+
+[tool.ruff.lint]
 select = [
     "E",    # pycodestyle errors
     "W",    # pycodestyle warnings
@@ -224,25 +226,25 @@ ignore = [
     "PLR0915", # too many statements
 ]
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "tests/*" = ["S101", "PLR2004", "ARG001", "ARG002"]
 "scripts/*" = ["INP001", "T201", "T203"]
 "examples/*" = ["INP001", "T201"]
 
-[tool.ruff.mccabe]
+[tool.ruff.lint.mccabe]
 max-complexity = 15
 
-[tool.ruff.pydocstyle]
+[tool.ruff.lint.pydocstyle]
 convention = "google"
 
-[tool.ruff.pylint]
+[tool.ruff.lint.pylint]
 max-args = 7
 max-branches = 15
 max-returns = 6
 max-statements = 50
 
 [tool.mypy]
-python_version = "3.9"
+python_version = "3.10"
 strict = true
 warn_return_any = true
 warn_unused_configs = true
