@@ -4,8 +4,6 @@ Support for Dockerfile language.
 
 from __future__ import annotations
 
-from typing import Optional
-
 from tree_sitter import Node
 
 from ..contracts.language_plugin_contract import ExtendedLanguagePluginContract
@@ -159,7 +157,7 @@ class DockerfilePlugin(LanguagePlugin, ExtendedLanguagePluginContract):
             return True
         return False
 
-    def get_node_context(self, node: Node, source: bytes) -> Optional[str]:
+    def get_node_context(self, node: Node, source: bytes) -> str | None:
         """Extract meaningful context for a node."""
         if node.type == "from_instruction":
             # Extract the base image name
