@@ -5,187 +5,188 @@ Tree-sitter Chunker - top-level package.
 __version__ = "1.0.0"
 
 __all__ = [
-    # Core functions
-    "get_parser",
-    "chunk_file",
-    "chunk_text_with_token_limit",
-    "chunk_file_with_token_limit",
-    "count_chunk_tokens",
-    # New parser API
-    "list_languages",
-    "get_language_info",
-    "return_parser",
-    "clear_cache",
-    # Configuration
-    "ParserConfig",
-    # Exceptions
-    "ChunkerError",
-    "LanguageNotFoundError",
-    "ParserError",
-    "LibraryNotFoundError",
-    # Performance features
-    "chunk_file_streaming",
-    "chunk_files_parallel",
-    "chunk_directory_parallel",
     "ASTCache",
-    "StreamingChunker",
-    "ParallelChunker",
-    "CodeChunk",
-    # Plugin system
-    "PluginManager",
-    "ChunkerConfig",
-    "LanguagePlugin",
-    "PluginConfig",
-    "get_plugin_manager",
-    # Enhanced chunking strategies
-    "SemanticChunker",
-    "HierarchicalChunker",
-    "AdaptiveChunker",
-    "CompositeChunker",
-    # Analysis tools
-    "ComplexityAnalyzer",
-    "CouplingAnalyzer",
-    "SemanticAnalyzer",
-    # Configuration system
-    "StrategyConfig",
-    "ChunkingProfile",
-    "get_profile",
-    "list_profiles",
     # Debug tools
     "ASTVisualizer",
-    "QueryDebugger",
-    "ChunkDebugger",
-    "NodeExplorer",
-    "start_repl",
-    "render_ast_graph",
-    "print_ast_tree",
-    "highlight_chunk_boundaries",
-    # Token counting integration
-    "TiktokenCounter",
-    "TokenAwareChunker",
-    "TreeSitterTokenAwareChunker",
-    # Hierarchy features
-    "ChunkHierarchyBuilder",
-    "HierarchyNavigator",
-    "ChunkHierarchy",
+    "AdaptiveChunker",
+    "AdvancedQueryIndex",
+    "AutoChunkResult",
+    "BaseCommentBlockRule",
+    "BaseComplexityAnalyzer",
     # Custom rules
     "BaseCustomRule",
-    "BaseRegexRule",
-    "BaseCommentBlockRule",
-    "MetadataRule",
-    "DefaultRuleEngine",
-    "TodoCommentRule",
-    "CopyrightHeaderRule",
-    "DocstringRule",
-    "ImportBlockRule",
-    "CustomMarkerRule",
-    "SectionHeaderRule",
-    "ConfigurationBlockRule",
-    "LanguageSpecificCommentRule",
-    "DebugStatementRule",
-    "TestAnnotationRule",
-    "get_builtin_rules",
     # Metadata extraction
     "BaseMetadataExtractor",
-    "BaseComplexityAnalyzer",
-    "PythonMetadataExtractor",
-    "PythonComplexityAnalyzer",
-    "JavaScriptMetadataExtractor",
-    "JavaScriptComplexityAnalyzer",
-    "SignatureInfo",
+    "BaseRegexRule",
+    "CacheEntry",
+    "ChangeDetector",
+    "ChangeType",
+    "ChunkBoundaryAnalyzer",
+    "ChunkCache",
+    "ChunkChange",
+    "ChunkDebugger",
+    "ChunkDiff",
+    "ChunkHierarchy",
+    # Hierarchy features
+    "ChunkHierarchyBuilder",
+    # Optimization (Phase 10)
+    "ChunkOptimizer",
+    # Advanced query (Phase 10)
+    "ChunkQueryAdvanced",
+    "ChunkerConfig",
+    # Exceptions
+    "ChunkerError",
+    "ChunkingProfile",
+    "CodeChunk",
+    # Analysis tools
+    "ComplexityAnalyzer",
     "ComplexityMetrics",
+    "CompositeChunker",
+    "CompositeFileSystem",
+    "ConfigurationBlockRule",
+    "ContextCache",
+    "ContextMetadata",
+    "ContextStrategy",
+    "CopyrightHeaderRule",
+    "CouplingAnalyzer",
+    "CrossLanguageReference",
+    "CustomMarkerRule",
+    "DebugStatementRule",
+    "DefaultChangeDetector",
+    "DefaultChunkCache",
+    "DefaultIncrementalProcessor",
+    "DefaultRuleEngine",
+    "DocstringRule",
+    "EmbeddedLanguageType",
+    "FileChunkResult",
+    # Garbage Collection tuning
+    "GCTuner",
+    "GitAwareProcessor",
+    "GitAwareProcessorImpl",
+    "GitignoreMatcher",
+    # Phase 19 components
+    "GrammarManager",
+    "HTTPFileSystem",
+    "HierarchicalChunker",
+    "HierarchyNavigator",
+    "HybridContextStrategy",
+    "ImportBlockRule",
+    "InMemoryContextCache",
+    "InMemoryFileSystem",
+    "IncrementalIndex",
+    # Incremental processing (Phase 10)
+    "IncrementalProcessor",
+    # Intelligent fallback
+    "IntelligentFallbackChunker",
+    "JavaScriptComplexityAnalyzer",
+    "JavaScriptMetadataExtractor",
+    "LanguageDetector",
+    "LanguageDetectorImpl",
+    "LanguageNotFoundError",
+    "LanguagePlugin",
+    "LanguageRegion",
+    "LanguageSpecificCommentRule",
+    "LibraryNotFoundError",
+    "LocalFileSystem",
+    "MemoryOptimizer",
+    "MergeConfig",
+    "MetadataRule",
+    # Multi-language processing
+    "MultiLanguageProcessor",
+    "MultiLanguageProcessorImpl",
+    "NaturalLanguageQueryEngine",
+    "NodeExplorer",
+    "ObjectPool",
+    "OptimizationConfig",
+    "OptimizationMetrics",
+    "OptimizationStrategy",
+    "OverlapConfig",
+    "OverlapStrategy",
+    # Overlapping fallback chunker
+    "OverlappingFallbackChunker",
+    "ParallelChunker",
+    # Configuration
+    "ParserConfig",
+    "ParserError",
+    "PluginConfig",
+    # Plugin system
+    "PluginManager",
+    "ProjectAnalyzer",
+    "ProjectAnalyzerImpl",
+    "PythonComplexityAnalyzer",
+    "PythonMetadataExtractor",
+    "QueryDebugger",
+    "QueryIndexAdvanced",
+    "QueryResult",
+    "QueryType",
+    "RelevanceContextStrategy",
+    "RepoChunkResult",
     # Repository processing
     "RepoProcessor",
-    "GitAwareProcessor",
     "RepoProcessorImpl",
-    "GitAwareProcessorImpl",
-    "FileChunkResult",
-    "RepoChunkResult",
-    "GitignoreMatcher",
-    "load_gitignore_patterns",
+    "SectionHeaderRule",
+    "SemanticAnalyzer",
+    # Enhanced chunking strategies
+    "SemanticChunker",
+    "SignatureInfo",
+    "SimpleIncrementalIndex",
+    # Smart context (Phase 10)
+    "SmartContextProvider",
+    "SmartQueryOptimizer",
+    # Configuration system
+    "StrategyConfig",
+    "StreamingChunker",
+    "TemplateGenerator",
+    "TestAnnotationRule",
+    # Token counting integration
+    "TiktokenCounter",
+    "TodoCommentRule",
+    "TokenAwareChunker",
     # Semantic merging
     "TreeSitterRelationshipAnalyzer",
     "TreeSitterSemanticMerger",
-    "MergeConfig",
-    # Overlapping fallback chunker
-    "OverlappingFallbackChunker",
-    "OverlapStrategy",
-    "OverlapConfig",
-    # Intelligent fallback
-    "IntelligentFallbackChunker",
-    # Smart context (Phase 10)
-    "SmartContextProvider",
     "TreeSitterSmartContextProvider",
-    "ContextMetadata",
-    "ContextStrategy",
-    "RelevanceContextStrategy",
-    "HybridContextStrategy",
-    "ContextCache",
-    "InMemoryContextCache",
-    # Advanced query (Phase 10)
-    "ChunkQueryAdvanced",
-    "NaturalLanguageQueryEngine",
-    "QueryIndexAdvanced",
-    "AdvancedQueryIndex",
-    "SmartQueryOptimizer",
-    "QueryType",
-    "QueryResult",
-    # Optimization (Phase 10)
-    "ChunkOptimizer",
-    "ChunkBoundaryAnalyzer",
-    "OptimizationMetrics",
-    "OptimizationConfig",
-    "OptimizationStrategy",
-    # Multi-language processing
-    "MultiLanguageProcessor",
-    "LanguageDetector",
-    "ProjectAnalyzer",
-    "LanguageRegion",
-    "CrossLanguageReference",
-    "EmbeddedLanguageType",
-    "MultiLanguageProcessorImpl",
-    "LanguageDetectorImpl",
-    "ProjectAnalyzerImpl",
-    # Incremental processing (Phase 10)
-    "IncrementalProcessor",
-    "ChunkCache",
-    "ChangeDetector",
-    "IncrementalIndex",
-    "ChunkChange",
-    "ChunkDiff",
-    "CacheEntry",
-    "ChangeType",
-    "DefaultIncrementalProcessor",
-    "DefaultChunkCache",
-    "DefaultChangeDetector",
-    "SimpleIncrementalIndex",
+    "TreeSitterTokenAwareChunker",
+    "VFSChunker",
+    "VirtualFile",
     # Virtual File System support
     "VirtualFileSystem",
-    "LocalFileSystem",
-    "InMemoryFileSystem",
-    "ZipFileSystem",
-    "HTTPFileSystem",
-    "CompositeFileSystem",
-    "VirtualFile",
-    "create_vfs",
-    "VFSChunker",
-    "chunk_from_url",
-    "chunk_from_zip",
-    # Garbage Collection tuning
-    "GCTuner",
-    "MemoryOptimizer",
-    "ObjectPool",
-    "get_memory_optimizer",
-    "tune_gc_for_batch",
-    "tune_gc_for_streaming",
-    "optimized_gc",
-    "gc_disabled",
     # Zero-configuration API (Phase 14)
     "ZeroConfigAPI",
-    "AutoChunkResult",
-    # Phase 19 components
-    "GrammarManager",
-    "TemplateGenerator",
+    "ZipFileSystem",
+    "chunk_directory_parallel",
+    "chunk_file",
+    # Performance features
+    "chunk_file_streaming",
+    "chunk_file_with_token_limit",
+    "chunk_files_parallel",
+    "chunk_from_url",
+    "chunk_from_zip",
+    "chunk_text",
+    "chunk_text_with_token_limit",
+    "clear_cache",
+    "count_chunk_tokens",
+    "create_vfs",
+    "gc_disabled",
+    "get_builtin_rules",
+    "get_language_info",
+    "get_memory_optimizer",
+    # Core functions
+    "get_parser",
+    "get_plugin_manager",
+    "get_profile",
+    "highlight_chunk_boundaries",
+    # New parser API
+    "list_languages",
+    "list_profiles",
+    "load_gitignore_patterns",
+    "optimized_gc",
+    "print_ast_tree",
+    "render_ast_graph",
+    "return_parser",
+    "start_repl",
+    "tune_gc_for_batch",
+    "tune_gc_for_streaming",
 ]
 
 # Analysis tools
@@ -197,6 +198,7 @@ from .cache import ASTCache
 from .chunker import (
     chunk_file,
     chunk_file_with_token_limit,
+    chunk_text,
     chunk_text_with_token_limit,
     count_chunk_tokens,
 )
