@@ -2,7 +2,9 @@
 
 from collections import defaultdict
 
-from ..interfaces.hierarchy import ChunkHierarchy
+from ..interfaces.hierarchy import (
+    ChunkHierarchy,
+)
 from ..interfaces.hierarchy import (
     ChunkHierarchyBuilder as ChunkHierarchyBuilderInterface,
 )
