@@ -13,8 +13,9 @@ import sys
 from pathlib import Path
 
 try:
-    from build import ProjectBuilder
     from build.env import IsolatedEnvBuilder
+
+    from build import ProjectBuilder
 except ImportError:
     print("Please install 'build' package: pip install build")
     sys.exit(1)
